Update the sensitivity tutorial
- use the SystematicSensitivity feature of Hades2
- overwrite config with JSON parameters
- match the corresponding new website page with an initial/complete structure
