+++ sensitivity/complete/pom.xml
@@ -3,26 +3,34 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+
     <parent>
-        <artifactId>powsybl-tutorials</artifactId>
-        <groupId>com.powsybl.tutorials</groupId>
-        <version>1.3.0-SNAPSHOT</version>
+        <groupId>com.powsybl</groupId>
+        <artifactId>powsybl-parent</artifactId>
+        <version>3</version>
+        <relativePath/>
     </parent>
 
     <artifactId>powsybl-sensitivity</artifactId>
-    <name>Sensitivity analysis tutorial</name>
 
-    <!-- These properties are only needed to run the tutorial easily -->
     <properties>
+        <maven.exec.version>1.6.0</maven.exec.version>
+        <slf4j.version>1.7.22</slf4j.version>
+        <powsybl.core.version>3.4.0</powsybl.core.version>
+        <powsybl.rte-core.version>2.11.0</powsybl.rte-core.version>
+
+        <!-- These properties are only needed to run the tutorial easily -->
         <exec.cleanupDaemonThreads>false</exec.cleanupDaemonThreads>
-        <exec.mainClass>powsybl.tutorials.sensitivity.SensitivityTutorial</exec.mainClass>
+        <exec.mainClass>powsybl.tutorials.sensitivity.SensitivityTutorialComplete</exec.mainClass>
     </properties>
 
+    <!-- Configure the tutorial to use a locally defined config file-->
     <build>
         <plugins>
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>exec-maven-plugin</artifactId>
+                <version>${maven.exec.version}</version>
                 <configuration>
                     <systemProperties>
                         <systemProperty>
@@ -37,16 +45,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.google.auto.service</groupId>
-            <artifactId>auto-service</artifactId>
-            <version>${autoservice.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <version>${guava.version}</version>
-        </dependency>
-        <dependency>
             <groupId>com.powsybl</groupId>
             <artifactId>powsybl-config-classic</artifactId>
             <version>${powsybl.core.version}</version>
@@ -55,16 +53,6 @@
             <groupId>com.rte-france.powsybl</groupId>
             <artifactId>powsybl-hades2-integration</artifactId>
             <version>${powsybl.rte-core.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.powsybl</groupId>
-            <artifactId>powsybl-iidm-impl</artifactId>
-            <version>${powsybl.core.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.powsybl</groupId>
-            <artifactId>powsybl-iidm-converter-api</artifactId>
-            <version>${powsybl.core.version}</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
