+++ sensitivity/complete/src/main/resources/config.yml
@@ -1,0 +1,25 @@
+loadflow:
+  default-impl-name: "hades2"
+
+computation-local:
+    tmp-dir: /tmp
+    availableCore: 1
+
+hades2:
+    homeDir: <PATH_TO_HADES2>
+    debug: true
+
+load-flow-default-parameters:
+    voltageInitMode: DC_VALUES
+    transformerVoltageControlOn: false
+    specificCompatibility: true
+
+hades2-default-parameters:
+    dcMode: false
+
+hades2-default-sensitivity-parameters:
+    computeInitialLoadflow: true
+    computeSensitivityToPsts: true
+    computeSensitivityToInjections: false
+    resultsThreshold: 0.01
+
