+++ sensitivity/initial/src/main/resources/sensi_parameters.json
@@ -1,0 +1,9 @@
+{
+  "extensions": {
+    "ADNSensitivityComputationParameters": {
+      "computeSensitivityToInjections": true,
+      "resultsThreshold": 0.1,
+      "hubPtdf": -1
+    }
+  }
+}
