--- sensitivity/src/main/resources/config.yml
@@ -1,29 +1,0 @@
-componentDefaultConfig:
-    LoadFlowFactory: com.rte_france.powsybl.hades2.Hades2Factory
-    SecurityAnalysisFactory: com.rte_france.powsybl.hades2.Hades2SecurityAnalysisFactory
-    ContingenciesProviderFactory: com.powsybl.action.dsl.GroovyDslContingenciesProviderFactory
-    
-#computation-local:
-#    tmp-dir: <path_to_local_tmp>
-#    availableCore: 1
-#
-#hades2:
-#    homeDir: <path_to_hades_2>
-#    debug: true
-#    useSAX: false
-
-#hades2-default-parameters:
-#    dcMode: true
-#
-#hades2-default-sensitivity-parameters:
-#    computeInitialLoadflow: true
-#    computeSensitivityToPsts: true
-#    computeSensitivityToInjections: true
-#    equationTypePst: ANGLE_SHIFT_EQ
-#    resultsThreshold: -1
-#    hubPtdf: -1
-#
-#load-flow-default-parameters:
-#    voltageInitMode: DC_VALUES
-#    transformerVoltageControlOn: false
-#    specificCompatibility: true
